pyproject.toml
[tool.poetry]
name = "fmcardgen"
version = "0.1.0"
description = ""
authors = ["Jacob Kaplan-Moss <jacob@jacobian.org>"]

[tool.poetry.scripts]
fmcardgen = 'fmcardgen.cli:cli'

[tool.pytest.ini_options]
addopts = "--cov=fmcardgen --no-cov-on-fail --cov-report term --cov-report html"
testpaths = "tests"

[tool.poetry.dependencies]
python = "^3.8"
typer = {extras = ["all"], version = "^0.3.2"}
Pillow = "^8.1.0"
python-frontmatter = "^0.5.0"
toml = "^0.10.2"
pydantic = "^1.7.3"
rich = "^9.11.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-cov = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
